move kafka error util func to kafka file
