+++ backend/pkg/api/connect/errors/kafka.go
@@ -11,6 +11,7 @@ package errors
 
 import (
 	"errors"
+	"strconv"
 
 	"connectrpc.com/connect"
 	"github.com/twmb/franz-go/pkg/kerr"
@@ -32,4 +33,30 @@ func NewConnectErrorFromKafkaErrorCode(code int16, msg *string) *connect.Error {
 		errors.New(errMsg),
 		NewErrorInfo(v1alpha1.Reason_REASON_KAFKA_API_ERROR.String(), KeyValsFromKafkaError(kafkaErr)...),
 	)
+}
+
+// KeyValsFromKafkaError tries to check if a given error is a Kafka error.
+// If this is the case, this function extracts the Kafka error code (int16)
+// as well as the string enum of this error code and returns a Key-Value
+// pair for each. These Key-Value pairs can be attached to the connect errors.
+func KeyValsFromKafkaError(err error) []KeyVal {
+	if err == nil {
+		return []KeyVal{}
+	}
+
+	var kafkaErr *kerr.Error
+	if errors.As(err, &kafkaErr) {
+		return []KeyVal{
+			{
+				Key:   "kafka_error_code",
+				Value: strconv.Itoa(int(kafkaErr.Code)),
+			},
+			{
+				Key:   "kafka_error_message",
+				Value: kafkaErr.Message,
+			},
+		}
+	}
+
+	return []KeyVal{}
 }
